requirements.txt
streamlit
requests
noisereduce>=0.3.0
librosa>=0.10.0
soundfile>=0.12.1
webrtcvad>=2.0.10
numpy>=1.20.0
scipy>=1.7.0
torch>=1.13.0
transformers>=4.25.1
asyncio>=3.4.3
aiohttp>=3.8.1
async-timeout>=4.0.2
streamlit-extras>=0.2.0
moviepy>=1.0.3
python-dotenv>=0.19.0